#4 reuse targetDogu and targetComponent for blueprint and effective blueprint
