+++ pkg/adapter/serializer/targetComponent.go
@@ -1,0 +1,52 @@
+package serializer
+
+import (
+	"errors"
+	"fmt"
+	"github.com/cloudogu/cesapp-lib/core"
+	"github.com/cloudogu/k8s-blueprint-operator/pkg/domain"
+)
+
+type TargetComponent struct {
+	// Name defines the name of the component including its namespace, f. i. "official/nginx". Must not be empty.
+	Name string `json:"name"`
+	// Version defines the version of the dogu that is to be installed. Must not be empty if the targetState is "present";
+	// otherwise it is optional and is not going to be interpreted.
+	Version string `json:"version"`
+	// TargetState defines a state of installation of this component. Optional field, but defaults to "TargetStatePresent"
+	TargetState string `json:"targetState"`
+}
+
+func ConvertComponents(components []TargetComponent) ([]domain.Component, error) {
+	var convertedComponents []domain.Component
+	var errorList []error
+
+	for _, component := range components {
+		newState, err := ToDomainTargetState(component.TargetState)
+		errorList = append(errorList, err)
+		if err != nil {
+			errorList = append(errorList, err)
+			continue
+		}
+		var version core.Version
+		if component.Version != "" {
+			version, err = core.ParseVersion(component.Version)
+			if err != nil {
+				errorList = append(errorList, fmt.Errorf("could not parse version of target component %q: %w", component.Name, err))
+				continue
+			}
+		}
+		convertedComponents = append(convertedComponents, domain.Component{
+			Name:        component.Name,
+			Version:     version,
+			TargetState: newState,
+		})
+	}
+
+	err := errors.Join(errorList...)
+	if err != nil {
+		return convertedComponents, fmt.Errorf("cannot convert blueprint components: %w", err)
+	}
+
+	return convertedComponents, err
+}
